requirements.txt
flask
opencv-python
tensorflow
Pillow
pillow_heif
gunicorn